html meta: twitter:card player => summary (*)
Using a player card sadly always embeds the iframe and plays the video when clicking on the card.

That breaks user experience. Users are only presented with an embedded video on the first click. They won't recognize that clicking a second time on
the card will bring them to the website.

I wrongly assumed it'd only play the video when clicking on the play icon on the image.