host: fix server version of commands
The listed commands were that of server version 1.0.3, because it contained already some commands that weren't there with 1.0.2.
(Because when I wrote it 1.0.2 was the latest release, but the master branch that I was looking at already contained additional commands.)

Maybe someday I'll add the info what command was added with which server version.