cc by-sa 4.0 license for files within /public/data/
All files in this repository are released under the MPL-2.0 license, including the /public/data/ directory.

This commits adds an extra license that is better suited for open data only to the /public/data/ directory and not to other parts of this repository.

By doing that, content inside /public/data/ is released under two licenses. The user can pick the license whiche suites their needs better (you don't
have to comply with both licenses simultanously).